refactor: Improve getAccountBalance and related functions
This commit improves the getAccountBalance function and related functions in the transactions.server.ts file. It introduces better handling of different account types (asset and liability) and calculates the account balance accordingly. Additionally, it adds the getAccountBalanceByType and getAccountBalanceByAccountId functions to retrieve account balances based on type and account ID, respectively. These changes enhance the accuracy and efficiency of retrieving account balances in the application.