feat: Add .dockerignore file and update npm dependencies
This commit adds a new .dockerignore file to the repository, which specifies the files and directories that should be ignored when building Docker images. It also updates the npm dependencies in the package.json file, including adding the "@flydotio/dockerfile" dependency. These changes improve the development and deployment processes by optimizing the Docker image build and ensuring the latest stable versions of the npm dependencies are used.