refactor: Add account relation to Transaction model
Add the `accountId` field and the `account` relation to the `Transaction` model in the `prisma/schema.prisma` file. This change allows transactions to be associated with specific accounts, providing more accurate and detailed financial tracking. The `accountId` field is a foreign key referencing the `id` field in the `Account` model. This commit enhances the functionality and usability of the application by improving the data model and enabling better analysis of transactions.