feat: Add support for transferring funds between accounts
This commit updates the `createTransaction` function in the `transactions.server.ts` file to add support for transferring funds between accounts. It introduces the `transfer` transaction type and implements the logic for creating two transactions - one for the source account and one for the destination account. The account balances are also updated accordingly. This enhancement improves the functionality of the application by allowing users to easily transfer funds between their accounts.