chore: Update npm dependencies
This commit updates the npm dependencies in the package.json and package-lock.json files. It adds the "zod" library as a new dependency with version "^3.23.8". This update ensures that the application has the latest version of the "zod" library, which is used for account validation and data schema definition. Keeping dependencies up to date is important for security, bug fixes, and compatibility with other libraries and frameworks.